multiupload fix obsluga bledow plikow

+++ src/E4u/Form/Element/FileUpload.php
@@ -52,28 +52,64 @@ class FileUpload extends Element
     }
 
     /**
-     * @param  array $file
-     * @return FileUpload
+     * @param  string[] $file
+     * @return bool
      */
     protected function checkFile($file)
     {
-        if (!empty($file)) {
-            if (empty($file['tmp_name']) || empty($file['name'])) {
-                $this->addError('Błąd ładowania pliku.');
-            }
-            elseif (!is_uploaded_file($file['tmp_name'])) {
-                $this->addError(sprintf('Błąd ładowania pliku: %s.', $file['name']));
-            }
-            elseif ($file['error'] != UPLOAD_ERR_OK) {
-                unlink($file['tmp_name']);
-                $this->addError(sprintf('Błąd ładowania pliku (%d): %s.', (int)$file['error'], $file['name']));
-            } elseif ($file['size'] > $this->maxSize) {
-                unlink($file['tmp_name']);
-                $this->addError(sprintf('Zbyt duży rozmiar pliku (max. %.2f MB): %s.', $this->maxSize / 1048576, $file['name']));
-            }
+        if (empty($file)) {
+            return true;
         }
 
-        return $this;
+        $error = $this->uploadErrorMessage($file['error']);
+        if (empty($error)) {
+            return true;
+        }
+
+        if (!empty($file['tmp_name']) && is_uploaded_file($file['tmp_name'])) {
+            unlink($file['tmp_name']);
+        }
+
+        $this->addError(sprintf('Błąd ładowania pliku (%d): %s.', $file['error'], $file['name']));
+        return false;
+    }
+
+    /**
+     * @param  int $error
+     * @return string
+     */
+    private function uploadErrorMessage($error)
+    {
+        switch ($file['error']) {
+
+            case UPLOAD_ERR_OK:
+                return null;
+
+            case UPLOAD_ERR_INI_SIZE:
+                return 'The uploaded file exceeds the upload_max_filesize directive in php.ini.';
+
+            case UPLOAD_ERR_FORM_SIZE:
+                return 'The uploaded file exceeds the MAX_FILE_SIZE directive that was specified in the HTML form.';
+
+            case UPLOAD_ERR_PARTIAL:
+                return 'The uploaded file was only partially uploaded.';
+
+            case UPLOAD_ERR_NO_FILE:
+                return 'No file was uploaded.';
+
+            case UPLOAD_ERR_NO_TMP_DIR:
+                return 'Missing a temporary folder.';
+
+            case UPLOAD_ERR_CANT_WRITE:
+                return 'Failed to write file to disk.';
+
+            case UPLOAD_ERR_EXTENSION:
+                return 'File upload stopped by extension.';
+
+            default:
+                return 'Unknown upload error.';
+
+        }
     }
 
     /**
